src/requirements.txt
python-telegram-bot==13.13
python-docx==1.1.0